Make DesiredApplyBundle::from_trusted_bundle pub(super) to desired_state
Previously the constructor was pub(crate), meaning any future in-crate
caller could bypass export_bundle_from_manifest and hand a raw
ConfigExportBundle (e.g. one loaded from a user-supplied JSON file via
read_config_import_bundle) into DesiredApplyBundle::from_trusted_bundle
and thence into apply_desired_state_changes.

Move DesiredApplyBundle from shared.rs into desired_state/apply_bundle.rs.
The struct is still pub(crate) so it can be named in function
signatures elsewhere, but from_trusted_bundle is now pub(super) —
callable only from within the desired_state module tree. The only
production caller is convert.rs::export_bundle_from_manifest, which
is a sibling module. config_import.rs can only produce a
DesiredApplyBundle by going through the typed manifest entry point,
making the non-interference invariant unforgeable intra-crate.
