+++ crates/defra-agent-cli/src/desired_state/convert.rs
@@ -184,7 +184,7 @@ pub(crate) fn manifest_from_export_bundle(
 pub(crate) fn export_bundle_from_manifest(
     manifest: &DesiredStateManifest,
     access_mode: &str,
-) -> Result<super::super::DesiredApplyBundle> {
+) -> Result<super::DesiredApplyBundle> {
     let bundle = super::super::ConfigExportBundle {
         format: super::super::CONFIG_EXPORT_FORMAT.to_string(),
         agent_did: manifest.agent_principal.agent_did.clone(),
@@ -222,9 +222,7 @@ pub(crate) fn export_bundle_from_manifest(
             .map(serde_json::to_value)
             .collect::<serde_json::Result<Vec<_>>>()?,
     };
-    Ok(super::super::DesiredApplyBundle::from_trusted_bundle(
-        bundle,
-    ))
+    Ok(super::DesiredApplyBundle::from_trusted_bundle(bundle))
 }
 
 pub(crate) fn desired_from_value<T>(value: &Value, allowed_fields: &[&str]) -> Result<T>
